Refactor(jiwon/refactoring): 폴더 구조 변경 #20
