--- app/src/models/services/Product/Product.js
@@ -1,9 +1,0 @@
-const ProductStorage = require("./ProductStorage");
-
-class Product {
-  constructor(req) {
-    this.body = req.body;
-  }
-}
-
-module.exports = Product;
