+++ app/src/services/Product/ProductService.js
@@ -1,0 +1,9 @@
+const ProductRepository = require("../../repository/Product/ProductRepository");
+
+class ProductService {
+  constructor(req) {
+    this.body = req.body;
+  }
+}
+
+module.exports = ProductService;
